Cleaning & fix add earning button
